+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Adds xml namespaces and schema
 - Adds HyphenationService to cache and speed up hyphenation
 - Adds validation and settings conversion services
+- Adds PDF constructor arguments to DocumentViewHelper
 - Fixes MultiColumnViewHelper columns not correctly initialized
 - Changes page.margins to page.margin
 - Removes classes EmptyFPDI, EmptyTCPDF and BithostTCPDF (replaced by BasePDF)